tea/radio/radio.go
package radio

import (
	"strings"

	tea "github.com/charmbracelet/bubbletea"
)

var _ tea.Model = &Model{}

type Model struct {
	availableChoices []string

	cursor int
	choice string
}

func NewRadioModel(availableChoices []string) *Model {
	return &Model{
		availableChoices: availableChoices,
		cursor:           0,
		choice:           "",
	}
}

func (m Model) Init() tea.Cmd {
	return nil
}

func (m Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
	switch msg := msg.(type) {
	case tea.KeyMsg:
		switch msg.String() {
		case "ctrl+c", "q", "esc":
			return m, tea.Quit

		case "enter":
			m.choice = m.availableChoices[m.cursor]
			return m, tea.Quit

		case "down", "j":
			m.cursor++
			if m.cursor >= len(m.availableChoices) {
				m.cursor = 0
			}

		case "up", "k":
			m.cursor--
			if m.cursor < 0 {
				m.cursor = len(m.availableChoices) - 1
			}
		}
	}

	return m, nil
}

func (m Model) View() string {
	s := strings.Builder{}
	s.WriteString("Which action would you like to do?\n\n")

	for i := 0; i < len(m.availableChoices); i++ {
		if m.cursor == i {
			s.WriteString("(•) ")
		} else {
			s.WriteString("( ) ")
		}
		s.WriteString(m.availableChoices[i])
		s.WriteString("\n")
	}
	s.WriteString("\n(press q to quit)\n")

	return s.String()
}
